refactor(init): extract prompt input
